feat(map): add marker type to results display
