+++ map/src/components/results.tsx
@@ -9,7 +9,7 @@ import { ContactDetails, MarkerInfo } from 'src/data/markers';
 import { button, buttonPrimary } from 'src/styling/mixins';
 
 import styled, { CLS_SCREEN_LG_ONLY } from '../styling';
-import Services from './services';
+import MarkerType from './marker-type';
 
 interface Props {
   className?: string;
@@ -134,9 +134,7 @@ const Results = (props: Props) => {
                 <div className="location">{result.loc.description}</div>
               )}
               <div className="name">{result.contentTitle}</div>
-              {result.type.type === 'org' && (
-                <Services services={result.type.services} />
-              )}
+              <MarkerType type={result.type} />
             </div>
           </div>
         ))}
@@ -147,9 +145,7 @@ const Results = (props: Props) => {
           {selectedResult.loc.description && (
             <div className="location">{selectedResult.loc.description}</div>
           )}
-          {selectedResult.type.type === 'org' && (
-            <Services services={selectedResult.type.services} />
-          )}
+          <MarkerType type={selectedResult.type} />
           {selectedResult.contentBody && (
             <div className="content">{selectedResult.contentBody}</div>
           )}
